openstack-client-server/requirements.txt
flask
python-openstackclient
python-glanceclient
python-neutronclient
python-cinderclient
sqlalchemy
PyMySQL